init edge_traits
