+++ edge/weighted_flow_edge.hpp
@@ -1,6 +1,7 @@
-#ifndef CPP_GL_FLOW_EDGE
-#define CPP_GL_FLOW_EDGE
+#ifndef CPP_GL_WEIGHTED_FLOW_EDGE
+#define CPP_GL_WEIGHTED_FLOW_EDGE
 
+#include <edge/edge_traits.hpp>
 #include <utility/types.hpp>
 
 
@@ -9,27 +10,32 @@ namespace gl {
 
 template <
     index_t vertex_key_t = std::size_t,
+    numerical_t weight_t = std::size_t,
     numerical_t flow_t = std::size_t,
     typename data_t = std::nullopt_t
 >
-struct flow_edge {
-    flow_edge() = default;
+struct weighted_flow_edge {
+    weighted_flow_edge() = default;
 
-    explicit flow_edge (
+    explicit weighted_flow_edge (
         const vertex_key_t& source, 
         const vertex_key_t& destination,
+        const weight_t& weight,
         const flow_t& flow,
         const flow_t& capacity,
         const data_t& data
     ) : 
         source(source), 
         destination(destination), 
+        weight(weight),
         flow(flow),
         capacity(capacity),
         data(data) 
     {}
 
-    explicit flow_edge (const flow_edge<vertex_key_t, flow_t, data_t>& other) : 
+    explicit weighted_flow_edge (
+        const weighted_flow_edge<vertex_key_t, weight_t, flow_t, data_t>& other
+    ) : 
         source(other.source), 
         destination(other.destination),
         flow(other.flow),
@@ -37,12 +43,12 @@ struct flow_edge {
         data(other.data)
     {}
 
-    ~flow_edge() = default;
+    ~weighted_flow_edge() = default;
 
     vertex_key_t source;
     vertex_key_t destination;
 
-    const bool weight = true;
+    weight_t weight;
     flow_t flow;
     flow_t capacity;
 
@@ -51,39 +57,44 @@ struct flow_edge {
 
 
 
-template <index_t vertex_key_t, numerical_t flow_t>
-struct flow_edge <vertex_key_t, flow_t, std::nullopt_t> {
-    flow_edge() = default;
+template <index_t vertex_key_t, numerical_t weight_t, numerical_t flow_t>
+struct weighted_flow_edge <vertex_key_t, weight_t, flow_t, std::nullopt_t> {
+    weighted_flow_edge() = default;
 
-    flow_edge (
+    weighted_flow_edge (
         const vertex_key_t& source, 
         const vertex_key_t& destination,
+        const weight_t& weight,
         const flow_t& flow,
         const flow_t& capacity
     ) : 
         source(source), 
         destination(destination),
+        weight(weight),
         flow(flow),
         capacity(capacity)
     {}
 
-    flow_edge (const flow_edge<vertex_key_t, flow_t>& other) : 
+    weighted_flow_edge (
+        const weighted_flow_edge<vertex_key_t, weight_t, flow_t>& other
+    ) : 
         source(other.source), 
         destination(other.destination),
+        weight(other.weight),
         flow(other.flow),
         capacity(other.capacity)
     {}
 
-    ~flow_edge() = default;
+    ~weighted_flow_edge() = default;
 
     vertex_key_t source;
     vertex_key_t destination;
 
-    const bool weight = true;
+    weight_t weight;
     flow_t flow;
     flow_t capacity;
 };
 
 } // namespace gl
 
-#endif // CPP_GL_FLOW_EDGE
+#endif // CPP_GL_WEIGHTED_FLOW_EDGE
