+++ edges.hpp
@@ -1,9 +1,10 @@
 #ifndef CPP_GL_EDGES
 #define CPP_GL_EDGES
 
-#include <edges/edge.hpp>
-#include <edges/weighted_edge.hpp>
-#include <edges/flow_edge.hpp>
-#include <edges/weighted_flow_edge.hpp>
+#include <edge/edge_traits.hpp>
+#include <edge/edge.hpp>
+#include <edge/weighted_edge.hpp>
+#include <edge/flow_edge.hpp>
+#include <edge/weighted_flow_edge.hpp>
 
 #endif // CPP_GL_EDGES
